Just Keep Swimming!
In sketch.js:
Made 200 fish “swim” from right to left, and more fish can be added
with mouse click.
Added update and display function.
Added Starting number of fish function.
Under setup function added 200 fish.
Under draw still included mouseIsPressed if statement.

In bubble.js:
Instead of radius added fish width and height.
They move at a faster pace along the x: (0, -5).
In the ‘if’ statement changed everything to this.x. Then, taking from
Daily 3, had fish reappear from the right edge of the canvas (width):
if (this.x < this.fishWidth * -1[this indicates that when the x = -20,
or when the fish completely disappears off the canvas]) this.x =
this.fishWidth + width[then make the fish reappear just off the edge of
the canvas at 620.];